chore(@e2e): replace scroll on overlay, fix kick/ban test
